processing/requirements.txt
fastapi
uvicorn
aiokafka
python-dotenv
minio